requirements.txt
Flask
mysql-connector-python
SQLAlchemy
python-dotenv
Flask-SQLAlchemy